requirements.txt
slackclient
Adafruit_GPIO
Adafruit_SSD1306
RPi.GPIO
PIL
